* add fluentd configuration * bind --api flag to auth.apiURL config * bind --fluentd flag to feedback.fluent.baseURL config * split batch_root.go from batch_validate.go
